pyproject.toml
[project]
name = "doclens"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.4.2",
    "dvc[s3]>=3.65.0",
    "hydra-core>=1.3.2",
    "pre-commit>=4.5.1",
    "protobuf>=6.33.2",
    "pytorch-lightning>=2.6.0",
    "ruff>=0.14.10",
    "sentencepiece>=0.2.1",
    "torch>=2.9.1",
    "torchmetrics>=1.8.2",
    "transformers>=4.57.3",
]

[project.scripts]
doclens-download = "doclens.cmd:cmd_download"
doclens-process = "doclens.cmd:cmd_process"
doclens-train = "doclens.cmd:cmd_train"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
